refactor(view): move Output to view package
